Create approvereview function

+++ controllers/reviewController.js
@@ -52,7 +52,15 @@ export function deletReview(req,res){
     
     const email = req.params.email;
 
-    Review.deleteOne({email:email}).then(
+    if(req.user == null){
+        res.status(401).json({
+            message : "Please login and try again"
+        });
+        return
+    }
+
+    if(req.user.role != "admin"){
+        Review.deleteOne({email:email}).then(
         ()=>{
             res.json({
                 message : "Review deleted successfully"
@@ -63,4 +71,60 @@ export function deletReview(req,res){
                     error:"Review deletion failed"
                 });
             });
+            return
+    }
+
+    if(req.user.role == "customer"){
+        if(req.user.email == email){
+        Review.deleteOne({email:email}).then(
+        ()=>{
+            res.json({
+                message : "Review deleted successfully"
+            });
+        }).catch(
+            ()=>{
+                res.status(500).json({
+                    error:"Review deletion failed"
+                });
+            });
+    }else{
+        res.status(403).json({
+            message : "You are not authorized to perfom this action"
+        });
+    }
+    }
+}
+
+
+export function approveReview(req,res){
+    const email = req.params.email;
+
+    if(req.user == null){
+        res.status(401).json({
+            message : "Please login and try again"
+        })
+    return
+    }
+
+    if(req.user.role == "admin"){
+        Review.updateOne({
+            email : email,
+        },
+        {
+            isApproved : true,
+        }
+    ).then(
+        ()=>{
+            res.json({
+                message : "Review approved successfully"
+            });
+        }).catch(
+            ()=>{
+                error : "Review approval failed"
+            });
+    }else{
+        res.status(403).json({
+            message : "You are not an admin.Only admins can approve the reviews"
+        });
+    }
 }
